Revert "enable namespace-wide SA secret injection to remove imagepullsecrets req"
This reverts commit a737a47e6c324e30ef26e051916e01df430e9020.
