+++ assets/charts/gitea-values.yaml
@@ -1,5 +1,7 @@
 persistence:
   storageClass: "###ZARF_STORAGE_CLASS###"
+imagePullSecrets:
+  - name: "zarf-registry"
 gitea:
   admin:
     username: "zarf-git-user"
