+++ cli/zarf.yaml
@@ -7,10 +7,22 @@ metadata:
 components:
   - name: baseline
     required: true
+    charts:
+      - name: docker-registry
+        url: https://helm.twun.io
+        version: 1.10.1
+        namespace: registry
+      - name: gatekeeper
+        url: https://repo1.dso.mil/platform-one/big-bang/apps/core/policy.git
+        version: 3.5.1-bb.10
+        gitPath: chart
+        namespace: gatekeeper
     manifests:
-      - name: podinfo
-        # This will be built on the package create side and deployed as a regular manifest on package deploy
-        kustomizations:
-          - github.com/stefanprodan/podinfo//kustomize
+      - name: test-manifests
+        files:
+        - ../assets/manifests/regsitry/configmap.yaml
+        - ../assets/manifests/traefik/traefik-tls.yaml
     images:
-      - ghcr.io/stefanprodan/podinfo:6.0.3
+      - registry1.dso.mil/ironbank/kiwigrid/k8s-sidecar:1.3.0
+    repos:
+      - https://repo1.dso.mil/platform-one/big-bang/apps/core/cluster-auditor.git@0.3.0-bb.2
