+++ examples/data-injection/manifests/data-injection.yaml
@@ -22,3 +22,5 @@ spec:
         limits:
           memory: "128Mi"
           cpu: "500m"
+  imagePullSecrets:
+    - name: zarf-registry
