+++ examples/postgres-operator/values/minio-operator.yaml
@@ -1,3 +1,5 @@
+imagePullSecrets:
+  - name: zarf-registry
 operator:
   image:
     repository: registry1.dso.mil/ironbank/opensource/minio/operator
