+++ examples/postgres-operator/values/postgres-operator-ui.yaml
@@ -1,3 +1,9 @@
+# image:
+  # registry: registry1.dso.mil
+  # repository: ??
+  # tag: ??
+imagePullSecrets:
+  - name: zarf-registry
 resources:
   requests:
     cpu: "100m"
