+++ examples/postgres-operator/values/postgres-operator.yaml
@@ -1,5 +1,12 @@
+# image:
+  # Eventually we'll need this to come from Iron Bank
+  # registry: registry1.dso.mil
+  # repository: ??
+  # tag: ??
 # configGeneral:
   # docker_image: registry1.dso.mil/.../spilo-13:2.1-p1
+imagePullSecrets:
+  - name: zarf-registry
 configPostgresPodResources:
   default_cpu_request: "100m"
   default_memory_request: "100Mi"
