+++ examples/single-big-bang-package/twistlock-values.yaml
@@ -1,0 +1,2 @@
+imagePullSecrets:
+  - name: zarf-registry
