+++ examples/single-big-bang-package/zarf.yaml
@@ -18,6 +18,8 @@ components:
         version: 0.0.6-bb.1
         namespace: twistlock
         gitPath: chart
+        valuesFiles:
+          - twistlock-values.yaml
     # https://umbrella-bigbang-releases.s3-us-gov-west-1.amazonaws.com/umbrella/1.14.0/images.txt
     images:
       - registry1.dso.mil/ironbank/twistlock/defender/defender:20.12.531
