+++ examples/tiny-kafka/charts/strimzi-values.yaml
@@ -1,3 +1,5 @@
+image:
+  imagePullSecrets: zarf-registry
 imageRegistryOverride: registry1.dso.mil
 imageRepositoryOverride: ironbank/opensource/strimzi
 watchNamespaces:
